+++ functions/index.js
@@ -71,12 +71,25 @@ exports.solver = functions.runWith(runtimeOpts).https.onRequest((req, res) => {
   }
 });
 
+function applyOffsets(data, entries) {
+  entries.forEach(customEntry => {
+    const explodedPath = customEntry.path.split('/');
+    explodedPath.shift();
+    let pointer = data;
+    for (let fragment of explodedPath.slice(0, -1)) {
+      pointer = pointer[fragment];
+    }
+    pointer[explodedPath[explodedPath.length - 1]] += customEntry.offset;
+  });
+  return data;
+}
+
 exports.updateList = functions.runWith(runtimeOpts).https.onCall((data, context) => {
   const listRef = firestore.collection('lists').doc(data.uid);
   return firestore.runTransaction(transaction => {
     return transaction.get(listRef).then(listDoc => {
       const list = listDoc.data();
-      const [standard, custom] = data.diff.reduce((acc, entry) => {
+      const [standard, custom] = JSON.parse(data.diff).reduce((acc, entry) => {
         if (entry.custom) {
           acc[1].push(entry);
         } else {
@@ -85,13 +98,28 @@ exports.updateList = functions.runWith(runtimeOpts).https.onCall((data, context)
         return acc;
       }, [[], []]);
       applyPatch(list, standard);
-      custom.forEach(customEntry => {
-        const explodedPath = customEntry.path.split('/');
-        explodedPath.shift();
-        list[explodedPath[0]][+explodedPath[1]][explodedPath[2]] += customEntry.offset;
-      });
+      applyOffsets(list, custom);
       transaction.update(listRef, list);
-      return Promise.resolve();
+    });
+  });
+});
+
+exports.updateInventory = functions.runWith(runtimeOpts).https.onCall((data, context) => {
+  const ref = firestore.collection('user-inventories').doc(data.uid);
+  return firestore.runTransaction(transaction => {
+    return transaction.get(ref).then(doc => {
+      const docData = doc.data();
+      const [standard, custom] = JSON.parse(data.diff).reduce((acc, entry) => {
+        if (entry.custom) {
+          acc[1].push(entry);
+        } else {
+          acc[0].push(entry);
+        }
+        return acc;
+      }, [[], []]);
+      applyPatch(docData, standard);
+      applyOffsets(docData, custom);
+      transaction.update(ref, docData);
     });
   });
 });